Initial draft of distillation mslearn readme
